+++ packages/expo-router/src/layouts/StackClient.tsx
@@ -18,9 +18,10 @@ import {
   type StackNavigationState,
   type StackRouterOptions,
 } from '../react-navigation/native';
-import { makePopAction, type NativeStackNavigationOptions } from '../react-navigation/native-stack';
+import type { NativeStackNavigationOptions } from '../react-navigation/native-stack';
 import type { NativeStackNavigationConfig } from '../react-navigation/native-stack/types';
-import { unstable_integrateWithRouter } from '../standard-navigation';
+import { makePopAction } from '../react-navigation/native-stack/utils/makePopAction';
+import { IsWithinNativeNavigator, unstable_integrateWithRouter } from '../standard-navigation';
 import { subscribePopToTopOnParentTabPress } from '../standard-navigation/subscribePopToTopOnParentTabPress';
 import { isChildOfType } from '../utils/children';
 import { Protected } from '../views/Protected';
@@ -100,7 +101,11 @@ const Stack = Object.assign(
       [props.children]
     );
 
-    return <RNStack {...props} children={rnChildren} screenOptions={screenOptions} />;
+    return (
+      <IsWithinNativeNavigator value>
+        <RNStack {...props} children={rnChildren} screenOptions={screenOptions} />
+      </IsWithinNativeNavigator>
+    );
   },
   {
     Screen: StackScreen,