+++ packages/expo-router/src/react-navigation/stack/types.tsx
@@ -1,6 +1,8 @@
 import type * as React from 'react';
 import type { Animated, ColorValue, StyleProp, TextStyle, ViewStyle } from 'react-native';
+import type { NavigatorDescriptor } from 'standard-navigation';
 
+import type { StandardNavigatorEmit } from '../../standard-navigation';
 import type {
   HeaderBackButton,
   HeaderBackButtonDisplayMode,
@@ -47,19 +49,7 @@ export type StackNavigationEventMap = {
   gestureCancel: { data: undefined };
 };
 
-export type StackViewEmit = (
-  event:
-    | {
-        type: 'transitionStart' | 'transitionEnd';
-        target?: string;
-        data: { closing: boolean };
-      }
-    | {
-        type: 'gestureStart' | 'gestureEnd' | 'gestureCancel';
-        target?: string;
-        data?: undefined;
-      }
-) => void;
+export type StackViewEmit = StandardNavigatorEmit<StackNavigationEventMap>;
 
 export type StackViewState = Pick<StackNavigationState<ParamListBase>, 'index' | 'routes'>;
 
@@ -133,8 +123,9 @@ export type Scene = {
   /**
    * Descriptor object for the screen.
    */
-  descriptor: Omit<StackDescriptor, 'options'> & {
-    options: Omit<StackDescriptor['options'], keyof SceneOptionsDefaults> & SceneOptionsDefaults;
+  descriptor: Omit<StackViewDescriptor, 'options'> & {
+    options: Omit<StackViewDescriptor['options'], keyof SceneOptionsDefaults> &
+      SceneOptionsDefaults;
   };
   /**
    * Animated nodes representing the progress of the animation.
@@ -304,6 +295,12 @@ export type StackDescriptor = Descriptor<
 >;
 
 export type StackDescriptorMap = Record<string, StackDescriptor>;
+
+export type StackViewDescriptor = NavigatorDescriptor<StackNavigationOptions> & {
+  navigation?: StackNavigationProp<ParamListBase>;
+};
+
+export type StackViewDescriptorMap = Record<string, StackViewDescriptor>;
 
 export type StackNavigationOptions = StackHeaderOptions &
   Partial<TransitionPreset> & {